Improved the automatic fullnode installer script (fixes #284)

+++ contrib/validators/automatic-fullnode-installer.sh
@@ -1,3 +1,6 @@
+#!/bin/bash
+shopt -s expand_aliases
+
 ####################################
 ## Variables
 ####################################
@@ -29,13 +32,13 @@ fi
 
 if [ -z "$DAEMON_NAME" ]; then
   {
-    echo " " >> ~/.profile
-    echo "# Setup Cosmovisor" >> ~/.profile
-    echo "export DAEMON_NAME=desmosd" >> ~/.profile
-    echo "export DAEMON_HOME=$HOME/.desmosd" >> ~/.profile
-    echo "export DAEMON_RESTART_AFTER_UPGRADE=on" >> ~/.profile
-    source ~/.profile
-  } &> /dev/null
+    echo " "
+    echo "# Setup Cosmovisor"
+    echo "export DAEMON_NAME=desmosd"
+    echo "export DAEMON_HOME=$HOME/.desmosd"
+    echo "export DAEMON_RESTART_AFTER_UPGRADE=on"
+  } >> ~/.profile
+  source ~/.profile
 fi
 
 echo "===> Completed environmental variables setup"
@@ -44,55 +47,84 @@ echo ""
 ####################################
 ## Setup Cosmovisor
 ####################################
-echo "===> Setting up Cosmovisor"
 
-echo "=====> Downloading Cosmovisor"
-# Download Cosmovisor
-{
-  git clone https://github.com/cosmos/cosmos-sdk.git ~/cosmos
-  cd ~/cosmos/cosmovisor
-  make cosmovisor
-  cp cosmovisor $GOBIN/cosmovisor
-  cd ~
-} &> /dev/null
+COSMOVISOR_FILE="$GOBIN/cosmovisor"
+if [ ! -f "$COSMOVISOR_FILE" ]; then
+  echo "===> Installing Cosmovisor"
 
-# Prepare Cosmovisor
-echo "=====> Installing up Cosmovisor"
-{
-  wget -O desmosd-cosmovisor.zip http://ipfs.io/ipfs/QmfVPHGPEimn7BKQo5JNeyiPtjbkYWqfnEvUnqfAVQapUe
-  sudo rm -r ~/.desmosd
-  mkdir -p ~/.desmosd
-  unzip desmosd-cosmovisor.zip -d ~/.desmosd
-} &> /dev/null
+  {
+    git clone https://github.com/cosmos/cosmos-sdk.git ~/cosmos
+    cd ~/cosmos/cosmovisor
+    make cosmovisor
+    cp cosmovisor -t "$GOBIN"
+    cd ~
+  } &> /dev/null
 
-echo "===> Completed Cosmovisor setup"
-echo ""
+  echo "===> Cosmovisor installed"
+  echo ""
+fi
+
 
 ####################################
 ## Setup Desmos
 ####################################
 echo "===> Setting up Desmos"
 
-# Setup desmosd to use Cosmovisor
+# Backup the priv validator key
+VALIDATOR_PRIV_KEY="$HOME/.desmosd/config/priv_validator_key.json"
+BACKUP_FILE="$HOME/priv_validator_key.json"
+if [ -f "$VALIDATOR_PRIV_KEY" ]; then
+  echo "====> Backing up the private validator key"
+  cp "$VALIDATOR_PRIV_KEY" "$BACKUP_FILE"
+fi
+
+# Delete the old ~/.desmosd folder
+DESMOSD_FOLDER="$HOME/.desmosd"
+if [ -d "$DESMOSD_FOLDER" ]; then
+  echo "====> Removing existing desmosd folder"
+  sudo rm -r ~/.desmosd
+fi
+
+# Delete the old ~/.desmoscli folder
+DESMOSCLI_FOLDER="$HOME/.desmoscli"
+if [ -d "$DESMOSCLI_FOLDER" ]; then
+  echo "====> Removing existing desmoscli folder"
+  sudo rm -r ~/.desmoscli
+fi
+
+# Configure Cosmovisor
+echo "====> Configuring Cosmovisor"
+echo "This might take a while..."
 {
+  wget -q --show-progress -O ~/desmosd-cosmovisor.zip http://ipfs.io/ipfs/QmfVPHGPEimn7BKQo5JNeyiPtjbkYWqfnEvUnqfAVQapUe 2>&1
+  unzip -o ~/desmosd-cosmovisor.zip -d ~/.desmosd
+  rm ~/desmosd-cosmovisor.zip
+
   alias desmosd=~/.desmosd/cosmovisor/current/bin/desmosd
   alias desmoscli=~/.desmosd/cosmovisor/current/bin/desmoscli
 } &> /dev/null
 
-# Setup the chain
-echo "=====> Initializing the chain"
+# Initialize the chain
+echo "====> Initializing a new chain"
 {
-  desmosd init $MONIKER
+  cosmovisor init "$MONIKER"
 } &> /dev/null
 
+# Restore the priv validator key
+if [ -f "$BACKUP_FILE" ]; then
+  echo "====> Restoring private validator key"
+  cp "$BACKUP_FILE" "$VALIDATOR_PRIV_KEY"
+  rm "$BACKUP_FILE"
+fi
+
 # Download the genesis file
-echo "=====> Downloading the genesis file"
+echo "====> Downloading the genesis file"
 {
   curl https://raw.githubusercontent.com/desmos-labs/morpheus/master/genesis.json -o $HOME/.desmosd/config/genesis.json
 } &> /dev/null
 
 # Setup the persistent peers
-echo "=====> Setting persistent peers"
+echo "====> Setting persistent peers"
 {
   sed -i -e 's/persistent_peers = ""/persistent_peers = "7fed5624ca577eb0333d3631b5e4f16ba1736979@54.180.98.75:26656,5077b7964d71d8758f7fc01cac01d0e2d55b8c18@18.196.238.210:26656,bdd98ec74fe56146f08e886239e52373f6821ce3@51.15.113.208:26656,e30d9bb713d17d1e4380b2e2a6df4b5c76c73eb1@34.212.106.82:26656"/g' ~/.desmosd/config/config.toml
 } &> /dev/null
@@ -104,9 +136,8 @@ echo ""
 ## Setup the service
 ####################################
 echo "===> Setting up Desmos service"
-
+FILE=/etc/systemd/system/desmosd.service
 {
-  FILE=/etc/systemd/system/desmosd.service
   sudo tee $FILE > /dev/null <<EOF
 [Unit]
 Description=Desmos full node watched by Cosmovisor
@@ -126,6 +157,7 @@ EOF
 } &> /dev/null
 
 echo "====> Starting Desmos service"
+echo ""
 {
   sudo systemctl daemon-reload
   sudo systemctl enable desmosd
